gras: linker stuff + moving around symbols and shit
Basically there were some issues on OSX linking,
because gras wasnt explicitly specified,
and was technically exposed by the gr wrappers.

As part of this commit, some symbols were moved into the wrappers,
and some of the interfaces which were in gras for backwards compat purposes,
The compile is still going, but this already addresses a few issues.

Now, I don't want to go copying every symbol over,
the real solution is to link in gras into whatever library.

Probably another commit to come...